% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hildareadR.R
\docType{package}
\name{hildareadR}
\alias{hildareadR}
\alias{hildareadR-package}
\title{hildareadR: A package to extract domain variables from the HILDA dataset}
\description{
This package provides an easy way to extract and combine variables from multiple HILDA
waves (a wave refers to each separate survey in a series of related surveys), using the \code{read_hilda} function.
}
\details{
HILDA is collected and maintained by the Melbourne Instutite and distributed via dataverse.
To apply for access to the HILDA data, visit the Australian Data Archive (https://dataverse.ada.edu.au/).
More information about the HILDA Survey dataset can be found at (https://melbourneinstitute.unimelb.edu.au/hilda)
}
\author{
Sara Kalucza & Sebastian Kalucza
}
